⚡ :: 다이어로그 다른 부분 눌럿을 때
